Changed /Conversation file structure
