+++ src/App.js
@@ -1,14 +1,18 @@
 import React, { Component } from 'react';
+import { Route } from 'react-router-dom';
+
 import miniData from './miniData';
 import presentData from './presentData';
-import { Route } from 'react-router-dom';
-import OptionsContainer from './components/optionsPage/OptionsContainer';
-import Home from './components/home/Home';
+
 import Header from './components/Header';
-import ConjugatorContainer from './components/conjugatorPage/ConjugatorContainer'
-import Conversation from './components/conversationPage/Conversation';
+import Home from './components/home/Home';
 import BeginnerPage from './components/home/Levels/BeginnerPage';
-import ConversationTwo from './components/conversationPage/ChatbotConversation/ConversationTwo'
+
+import ConversationOne from './components/Conversation/ConversationOne';
+import ConversationTwo from './components/Conversation/ConversationTwo'
+
+import OptionsContainer from './components/optionsPage/OptionsContainer';
+import ConjugatorContainer from './components/conjugatorPage/ConjugatorContainer'
 
 class App extends Component {
   constructor() {
@@ -42,7 +46,7 @@ class App extends Component {
       <div style={{width: '420px', margin: '0 auto'}}>
         <Route path="/" component={Header} />
         <Route exact path="/" component={Home} />
-        <Route exact path="/conversation" component={Conversation} />
+        <Route exact path="/conversation" component={ConversationOne} />
         <Route exact path="/conversation/two" component={ConversationTwo} />
         <Route exact path="/beginner" component={BeginnerPage} />
         <Route
